requirements.txt
# Core dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.0
hume==0.7.0

# Optional: Google Cloud Storage (comment out if not needed)
google-cloud-storage==2.18.2
google-auth==2.35.0
